+++ package.json
@@ -747,6 +747,7 @@
 		"ms-rest-azure": "^2.3.1",
 		"opn": "^5.3.0",
 		"vscode-azureextensionui": "^0.28.3",
+		"vscode-nls": "^4.1.1",
 		"winreg": "^1.2.3"
 	}
 }
